Unban zefra
"Hi. I add you to make an unban request. a day ago I went in with a
friend to do a stupid troll to a player on a serveme.tf server. it was a
bad joke and we got banned. I play from time to time on tf2 center and I
can't get on the server to play because of the ban. I didn't know where
I could appeal for an unban and I'm writing to you here, and if i can
get help, next time it will not happen again."